refactor: move listing page as first level
