+++ packages/manager/modules/sharepoint/src/dashboard/sharepoint-order.service.js
