+++ packages/manager/modules/sharepoint/src/dashboard/sharepoint-tabs.controller.js
