+++ packages/manager/modules/sharepoint/src/dashboard/sharepoint.controller.js
