+++ packages/manager/modules/sharepoint/src/dashboard/sharepoint.html
