+++ packages/manager/modules/sharepoint/src/dashboard/sharepoint.service.js
