+++ packages/manager/modules/sharepoint/src/dashboard/translations/Messages_es_ES.json
