+++ packages/manager/modules/sharepoint/src/dashboard/translations/Messages_fr_CA.json
