+++ packages/manager/modules/sharepoint/src/dashboard/translations/Messages_pt_PT.json
