+++ packages/manager/modules/sharepoint/src/sharepoint.routing.js
