--- packages/manager/modules/sharepoint/src/sharepoints/sharepoints.module.js
@@ -1,21 +1,0 @@
-import angular from 'angular';
-import 'angular-translate';
-import '@ovh-ux/ui-kit';
-import ngTranslateAsyncLoader from '@ovh-ux/ng-translate-async-loader';
-import { ListLayoutHelper } from '@ovh-ux/manager-ng-layout-helpers';
-
-import routing from './sharepoints.routing';
-
-const moduleName = 'ovhManagerExchanges';
-
-angular
-  .module(moduleName, [
-    ngTranslateAsyncLoader,
-    'oui',
-    'pascalprecht.translate',
-    ListLayoutHelper.moduleName,
-  ])
-  .config(routing)
-  .run(/* @ngTranslationsInject:json ./translations */);
-
-export default moduleName;
