--- packages/manager/modules/sharepoint/src/sharepoints/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "sharepoint_title": "SharePoint"
-}
